Move tests to trait
